% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencoder}
\alias{autoencoder}
\title{Learn a deep autoencoder}
\usage{
autoencoder(x, hidden = c(150, 150, 500, 4), activation = c("softsign",
  "softsign", "softsign", "linear"), epochs = 100)
}
\arguments{
\item{x}{Numeric expression matrix}

\item{hidden}{Vector of neurons in the hidden layer(s)}

\item{activation}{Vector of hidden layer activations}

\item{epochs}{Number of training epochs}
}
\description{
Learn a deep autoencoder
}
