% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_storage.R
\name{push_fcs}
\alias{push_fcs}
\title{Upload file into local rsqlite database}
\usage{
push_fcs(path, name, sqlpath = "~/.deepflow/deepflow.db")
}
\arguments{
\item{path}{character string}

\item{name}{character string indicating table name}
}
\description{
Upload file into local rsqlite database
}
